Refactor monkey patch "PatchZ2LogTimezone".
