+++ opengever/base/monkey/patches/__init__.py
@@ -1,0 +1,4 @@
+from .tz_for_log import PatchZ2LogTimezone
+
+
+PatchZ2LogTimezone()()
